backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic-settings>=2.2.1

langchain>=1.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq==0.1.5
langgraph>=0.2.0
groq==0.9.0

httpx==0.27.0
requests==2.32.3

faiss-cpu==1.13.2
sentence-transformers==2.7.0